requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-json-widget==1.1.1
django-markdownx==3.0.1
django-modeltranslation==0.17.5
djangorestframework==3.13.1
future==0.18.2
idna==3.3
importlib-metadata==4.10.1
kombu==5.2.3
Markdown==3.3.6
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
pytz==2021.3
requests==2.27.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
stripe==2.66.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
